# Mutate newDirtyPaths instead of creating a new array every iteration