@zhujinxuan 2019-01-30
Hi, can I know how do you test it?

I remember both push.apply and concat have similar speed. (https://jsperf.com/array-concat-vs-push-2/17)  

But perhaps there is another optimization like this:
```js
    const dirty = Array.prototype.concat.apply(
        newDirtyPath, 
        this.tmp.dirty.map(path => {      
           path = PathUtils.create(path)
          return PathUtils.transform(path, operation).toArray()
     })
   );
```