@brendancarney 2019-02-01
Unfortunately, I think I spoke too soon. The test is slower in my actual slate example, so building the large array is likely slow. I'm concerned that different cases will have different results, so I'm going to keep experimenting and try to find a more reliable optimization here.