@zhujinxuan 2019-02-01
@brendancarney Can I have a loook at the benchmark result by:

1. in master branch, `yarn build && yarn benchmark:save`
2. in your branch, `yarn benchmark`