@brendancarney 2019-02-18
Sorry for the delay, I was out of town for a week. Unfortunately, I'm not sure there's a benchmark accurately reflecting my use case. Running the benchmarks show no meaningful improvements in my branch. 

However, I took a recent production example of a large document and tried pasting it into an editor when running the slate examples. Over several attempts with both branches, here are the rough averages:
master: ~60s
this PR: ~18s

Still really slow in both cases unfortunately. I understand that we'll probably want a benchmark that shows this improvement, so I'll try to get that added soon.