@brendancarney 2019-02-18
Ok so I sort of have a benchmark ready, but still trying to figure out the best way to introduce it. Currently, we run each benchmark a minimum of 100 times. This is a little problematic for this specific case because it is least performant with larger documents, and it's really really slow. So running it one hundred times is probably not great. Should we make these options configurable by benchmark? 

Here are the results from inserting a paragraph block 1,000 times, wrapped in `editor.withoutNormalizing`:

<img width="479" alt="screen shot 2019-02-18 at 3 31 05 pm" src="https://user-images.githubusercontent.com/2984336/52977360-92a46580-3392-11e9-9d7b-efd4b364abe0.png">

```
★ insert-block:
     user: 0.02 -> 0.07 ops/sec (217.23% faster 🙌 )
     real: 0.02 -> 0.08 ops/sec (230.25% faster 🙌 )
```
As you can see, still really really slow, but there is an improvement.