@zhujinxuan 2019-02-19
@brendancarney The benchmark is very naive at this moment.  But I think the 200% is very convincing.

If you are interested, you can temporarily add a large document insert benchmark inside `benchmark/slate/changes` and then run it with only `yarn benchmark:save --grep pattern` and  `yarn benchmark --grep pattern` (where pattern shall be a RegExp).  

BTW, can you have a try on the `Array.prototype.concat.apply` version in the paste large text example? (Perhaps not faster because map is slow...)

I am a bit surprising that the paste took 18 seconds... Last year I did same test after #1661, it was taking 5 seconds (though still slow).  I think I may need to investigate why when I have time.