@brendancarney 2019-02-21
Hi @zhujinxuan. It turns out they are about the same. I think your suggestion is slightly more readable. Thoughts?

Also, I did not commit a benchmark because it would be ridiculously slow with the current setup. 

To test the speed difference, you can cut and paste the large document example in master, and then in here.