@zhujinxuan 2019-02-23
@brendancarney As there is significant difference between `concat <<< map` and push, I would suggest keep the push version.